smart_parking_backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
flask==2.2.3
flask-jwt-extended==4.4.4
firebase-admin==6.1.0
geopy==2.3.0
python-dateutil==2.8.2
bcrypt==4.0.1
email-validator==2.0.0
pydantic==1.10.7